fix(example-runner): do not override kendo-ui & jQuery packages
Add mapping for jQuery only when the module is part of the auto imports.

Add mapping and package configuration for Kendo UI only when it is part of the auto import. Allow more granular configuration
based on the main file field. If no main file is provided fallback for kendo.all.js.