chore: remove unused templates, simplify post-processing of templates
with the new inferral of snippet types, the multiple file types in the templates are not needed. this allows us to remove duplicate templates, and thus remove the filtering code from snippets.js
also, the two $.each loops over the same object can be merged by moving the conditions inside. Generally these conditions should only touch the editor templates _or_ be in prepareSnippet, which is the next step